producer3/requirements.txt
requests
pymongo
kafka-python
requests
